+++ docs/en/docs/release-notes.md
@@ -10,6 +10,7 @@ hide:
 ### Added
 
 - Support for automatic cast call for a function is a function is provided to cast in the `loader` of the [Environment](./environments.md).
+- `UUIDEncoder` as a native for parsing `UUID` types.
 
 ### Fixed